Move `KeyboardShortcut` into a separete file
