+++ src/components/views/settings/tabs/user/KeyboardUserSettingsTab.tsx
@@ -18,63 +18,16 @@ limitations under the License.
 import React from "react";
 
 import {
-    ALTERNATE_KEY_NAME,
-    KEY_ICON,
     ICategory,
     CATEGORIES,
     CategoryName,
 } from "../../../../../accessibility/KeyboardShortcuts";
 import SdkConfig from "../../../../../SdkConfig";
-import { isMac, Key } from "../../../../../Keyboard";
 import { _t } from "../../../../../languageHandler";
 import {
-    getKeyboardShortcutDisplayName,
-    getKeyboardShortcutValue,
+    getKeyboardShortcutDisplayName, getKeyboardShortcutValue,
 } from "../../../../../accessibility/KeyboardShortcutUtils";
-
-interface IKeyboardKeyProps {
-    name: string;
-    last?: boolean;
-}
-
-export const KeyboardKey: React.FC<IKeyboardKeyProps> = ({ name, last }) => {
-    const icon = KEY_ICON[name];
-    const alternateName = ALTERNATE_KEY_NAME[name];
-
-    return <React.Fragment>
-        <kbd> { icon || (alternateName && _t(alternateName)) || name } </kbd>
-        { !last && "+" }
-    </React.Fragment>;
-};
-
-interface IKeyboardShortcutProps {
-    name: string;
-}
-
-export const KeyboardShortcut: React.FC<IKeyboardShortcutProps> = ({ name }) => {
-    const value = getKeyboardShortcutValue(name);
-    if (!value) return null;
-
-    const modifiersElement = [];
-    if (value.ctrlOrCmdKey) {
-        modifiersElement.push(<KeyboardKey key="ctrlOrCmdKey" name={isMac ? Key.META : Key.CONTROL} />);
-    } else if (value.ctrlKey) {
-        modifiersElement.push(<KeyboardKey key="ctrlKey" name={Key.CONTROL} />);
-    } else if (value.metaKey) {
-        modifiersElement.push(<KeyboardKey key="metaKey" name={Key.META} />);
-    }
-    if (value.altKey) {
-        modifiersElement.push(<KeyboardKey key="altKey" name={Key.ALT} />);
-    }
-    if (value.shiftKey) {
-        modifiersElement.push(<KeyboardKey key="shiftKey" name={Key.SHIFT} />);
-    }
-
-    return <div>
-        { modifiersElement }
-        <KeyboardKey name={value.key} last />
-    </div>;
-};
+import { KeyboardShortcut } from "../../KeyboardShortcut";
 
 interface IKeyboardShortcutRowProps {
     name: string;
@@ -86,11 +39,12 @@ const visibleCategories = Object.entries(CATEGORIES).filter(([categoryName]) =>
 
 const KeyboardShortcutRow: React.FC<IKeyboardShortcutRowProps> = ({ name }) => {
     const displayName = getKeyboardShortcutDisplayName(name);
-    if (!displayName) return null;
+    const value = getKeyboardShortcutValue(name);
+    if (!displayName || !value) return null;
 
     return <div className="mx_KeyboardShortcut_shortcutRow">
         { displayName }
-        <KeyboardShortcut name={name} />
+        <KeyboardShortcut value={value} />
     </div>;
 };
